Make Popup a wrapper for the Chat component
